add new sorter for sorting `SearchResult`
Sort by relevance is moved there and completely removed from `FileListSorter`
